+++ templates/partials/menu/mobile.php
@@ -1,7 +1,7 @@
 <?php namespace ProcessWire; ?>
 
-<input type="checkbox" id="mobile-menu-toggle" class="menu-toggle" aria-controls="mobile-menu" hidden>
-<label for="mobile-menu-toggle" class="menu-toggle-label" tabindex="0">
+<input type="checkbox" id="mobile-menu-toggle" class="toggle" aria-controls="mobile-menu" hidden>
+<label for="mobile-menu-toggle" class="toggle-label" tabindex="0">
     <i class="fas fa-bars" aria-hidden="true"></i>
     <span class="visually-hidden"><?= __('Toggle mobile menu') ?></span>
 </label>
@@ -14,5 +14,8 @@ echo $menu->render([
     'placeholders' => [
         'menu_class_modifier' => 'mobile',
         'aria_label' => 'Navigation',
+    ],
+    'classes' => [
+        'nav' => 'menu toggle-target',
     ],
 ]);