+++ templates/wireframe.php
@@ -19,11 +19,20 @@ $wireframe->init();
 require_once 'lib/config.php';
 
 // render the page
+$search_page = $pages->get('template=search');
 echo $wireframe->render([
     'site_name' => 'Lorem Ipsum',
     'lang' => 'en',
     'home' => $pages->get(1),
     'menu' => $modules->get('MarkupMenu'),
+    'search_form' => $search_page->id ? $modules->get('SearchEngine')->renderForm([
+        'form_action' => $search_page->url,
+        'form_id' => 'se-masthead-form',
+        'form_input_id' => 'se-masthead-form-input',
+        'classes' => [
+            'form' => 'search-form &--masthead',
+        ],
+    ]) : null,
     'body_class' => implode(array_filter([
         'template-' . ($wire->view->template ?: $page->template),
         $wire->view->template ? 'delegated-template-' . $page->template : '',